src/chat/requirements.txt
chainlit>=2.0.1
python-dotenv>=1.0.1
dapr>=1.14.0
semantic-kernel==1.20.0